% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesboot.R
\name{as.bayesboot}
\alias{as.bayesboot}
\title{Convert an object to a bayesboot object}
\usage{
as.bayesboot(object)
}
\arguments{
\item{object}{Any object that can be converted to a data frame.}
}
\value{
A \code{data.frame} with subclass \code{bayesboot}.
}
\description{
This converts an object into a dataframe and adds the class \code{bayesboot}.
Doing this is only usful in the case you would want to use the plot and summary methods
for \code{bayesboot} objects.
}

