% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesboot.R
\name{bayesboot}
\alias{bayesboot}
\title{Bayesian bootstrap}
\usage{
bayesboot(
  data,
  f = mean,
  b = 10000,
  plot = TRUE,
  quantilepoints = c(0.025, 0.975)
)
}
\arguments{
\item{data}{Data}

\item{f}{A function to estimate with bootstrap}

\item{b}{Bootstrap sample size}

\item{plot}{Should the results be plotted (default = TRUE)}

\item{quantilepoints}{Quantile points to be plotted (default = c(0.025, 0.975))}
}
\value{
Values of the the function 'f' calculated on the bayesian bootstrap data
}
\description{
Calculates the value of a function on the bayesian bootstrap data. Uses the dirichlet distribution to generate positive real valued weights for the data as opposed to using integer weights (normal bootstrapo).
}
\examples{
# Calculate bootstrap means
data = c(1, 1, 3, 4, 7, 12)
bayesboot(data)

}
