% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesboot.R
\name{bootstrap_bayes}
\alias{bootstrap_bayes}
\title{Bayesian bootstrap}
\usage{
bootstrap_bayes(
  data,
  f = mean,
  b = 10000,
  plot = TRUE,
  quantilepoints = c(0.025, 0.975)
)
}
\arguments{
\item{data}{Data}

\item{f}{A function to estimate with bootstrap}

\item{b}{Bootstrap sample size}

\item{plot}{Should the results be plotted (default = TRUE)}
}
\value{
Values of the the function 'f' calculated on the bayesian bootstrap data
}
\description{
Calculates the value of a function on the bayesian bootstrap data
}
\examples{
# Calculate bootstdocrap means
data = c(1, 1, 3, 4, 7, 12)
bootstrap_bayes(data)

}
