% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesboot.R
\name{plot.bayesboot}
\alias{plot.bayesboot}
\title{Plot the result of \code{bayesboot}}
\usage{
\method{plot}{bayesboot}(x, cred.mass = 0.95, plots.per.page = 3,
  cex = 1.2, cex.lab = 1.3, ...)
}
\arguments{
\item{x}{The bayesboot object to plot.}

\item{cred.mass}{the probability mass to include in credible intervals, or
NULL to suppress plotting of credible intervals.}

\item{plots.per.page}{The maximum numbers of plots per page.}

\item{cex, cex.lab, ...}{Further parameters passed on to
\code{\link{plotPost}}.}
}
\description{
Produces histograms showing the marginal posterior distributions from a
\code{bayesboot} call. Uses the \code{\link{plotPost}} function to produce
the individual histograms.
}

