% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesboot.R
\name{print.bayesboot}
\alias{print.bayesboot}
\title{Print the first number of draws from the Bayesian bootstrap}
\usage{
\method{print}{bayesboot}(x, n = 10, ...)
}
\arguments{
\item{x}{The bayesboot object to print.}

\item{n}{The number of draws to print.}

\item{...}{Not used.}
}
\description{
Print the first number of draws from the Bayesian bootstrap
}
