% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesboot.R
\name{rudirichlet}
\alias{rudirichlet}
\title{Produce random draws from a uniform Dirichlet distribution}
\usage{
rudirichlet(n, d)
}
\arguments{
\item{n}{the number of draws.}

\item{d}{the dimension of the Dirichlet distribution.}
}
\value{
An \code{n} by \code{d} matrix.
}
\description{
\code{rudirichlet} produces \code{n} draws from a \code{d}-dimensional uniform Dirichlet distribution.
Here "uniform" implies that any combination of values on the support of the distribution is equally likely,
that is, the \eqn{\alpha} parameters to the Dirichlet distribution are all set to 1.0.
}
\examples{
rudirichlet(2, 3)
# Should produce the following matrix, but with different numbers, of course:
#        [,1]     [,2]   [,3]
# [1,] 0.7109 0.001012 0.2881
# [2,] 0.2112 0.491784 0.2970

}

