% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesboot.R
\name{bayesboot}
\alias{bayesboot}
\title{bayes_boot}
\usage{
bayesboot(data, statistic, n1 = 1000, n2 = 1000, use_weights = FALSE,
  weight_arg = NULL, ...)
}
\arguments{
\item{...}{}
}
\description{
bayes_boot
}

